Mejorados los ejercicios y agregado el recuperatorio del primer parcial

+++ PJ_1_Scanner_Y_JOption_Pane_09052024/Ejercicio4_Comparacion.java
@@ -1,13 +1,13 @@
-/*Declara 2 variables numéricas (con el valor que desees), he indica cual es mayor de los dos. Si
+/*Declara 2 variables numéricas (con el valor que desees), e indica cual es mayor de los dos. Si
 son iguales indicarlo también. Ves cambiando los valores para comprobar que funciona. */
 import java.util.Scanner;
 public class Ejercicio4_Comparacion {
     public static void main(String[] args) {
         Scanner sc = new Scanner(System.in);
         System.out.println("Porfavor ingrese un número : ");
-        int numero1 = sc.nextInt();
+        double numero1 = sc.nextDouble();
         System.out.println("Porfavor ingrese otro número : ");
-        int numero2 = sc.nextInt();
+        double numero2 = sc.nextDouble();
         if (numero1 > numero2){
             System.out.println("El primer número es mayor que el segundo.");
         }else if(numero1 < numero2){
